[doom] fix auth-source credential loading (again)
Hopefully this is the last time!

It was causing non-deterministic startup crashes when the `(after!)`
hook ran at the wrong time. Something about the "git-global-commit-hook"
failing because the returned :secret was nil in doom-first-file-hook.
Blech.

I stress tested this solution and it seemed to work very consistently,
decrypting and loading the wakatime API key when opening the first file.
doom-first-file-hook didn't work for restarts or session restoring, but
this seems to, without creating startup errors.